pyproject.toml
[tool.poetry]
name = "tsocrapi"
version = "0.1.0"
description = "OCR API for TimeSlicer"
authors = ["navanchauhan <navanchauhan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.27.0"
fastapi = "^0.92.0"
azure-cognitiveservices-vision-computervision = "^0.9.0"
uvicorn = "^0.20.0"
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
